Hacking Framework/security_platform/requirements.txt
# ─── Core Framework ────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.8.0
pydantic-settings>=2.3.0

# ─── Database ──────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.31
# asyncpg: PostgreSQL async driver (optional — only needed with PostgreSQL)
asyncpg>=0.29.0
# aiosqlite: SQLite async driver (used in dev mode)
aiosqlite>=0.20.0
alembic>=1.13.0
# psycopg2-binary omitted: requires pg_config; install manually if using Alembic sync migrations

# ─── Cache & Task Queue ────────────────────────────────────────────────────
redis>=5.0.0
celery>=5.4.0
# flower>=2.0.1  # optional monitoring UI — comment out if install fails
kombu>=5.3.0

# ─── Authentication ────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9

# ─── Networking & Security Research ───────────────────────────────────────
dnspython>=2.6.0
httpx>=0.27.0
websockets>=12.0

# ─── CLI ───────────────────────────────────────────────────────────────────
typer[all]>=0.12.0
rich>=13.7.0

# ─── Reporting & Templating ────────────────────────────────────────────────
jinja2>=3.1.0
aiofiles>=24.1.0

# ─── Optional: Elasticsearch ───────────────────────────────────────────────
# elasticsearch[async]>=8.14.0  # only needed if ES_ENABLED=true

# ─── Utilities ─────────────────────────────────────────────────────────────
python-dateutil>=2.9.0
tenacity>=8.5.0

# ─── Testing ───────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
